requirements.txt
python-telegram-bot==20.7
stripe
python-dotenv
flask

